Combined ssnl unit tests
Moved all seasonal unit tests to the better named file.
Added more median1D tests.
Made PEP8 changes
Added docstrings
Removed unnecessary imports
